+++ lib/sprout/src/main/kotlin/org/partiql/sprout/generator/target/kotlin/KotlinOptions.kt
@@ -8,19 +8,4 @@ package org.partiql.sprout.generator.target.kotlin
 class KotlinOptions(
     val packageRoot: String,
     val poems: List<String>,
-    val node: KotlinNodeOptions = KotlinNodeOptions(),
 )
-
-/**
- * Consider other options as this is Kotlin specific
- */
-class KotlinNodeOptions(
-    val modifier: Modifier = Modifier.FINAL,
-) {
-
-    enum class Modifier {
-        FINAL,
-        DATA,
-        OPEN,
-    }
-}